Prepare to uniform the hvl_tables.c and hvl_tables.h files, by first making all of the replayers use them
