+++ Replayer_Amiga/hvl_replay.c
@@ -25,188 +25,9 @@
 #include "hvl_replay.h"
 #include "hvl_tables.h"
 
-
-/*
-** Waves
-*/
-
-
 int8 waves[WAVES_SIZE];
 
 uint32 panning_left[256], panning_right[256];
-
-void hvl_GenPanningTables( void )
-{
-  uint32 i;
-  float64 aa, ab;
-
-  // Sine based panning table
-  aa = (3.14159265f*2.0f)/4.0f;   // Quarter of the way through the sinewave == top peak
-  ab = 0.0f;                      // Start of the climb from zero
-
-  for( i=0; i<256; i++ )
-  {
-    panning_left[i]  = (uint32)(sin(aa)*255.0f);
-    panning_right[i] = (uint32)(sin(ab)*255.0f);
-    
-    aa += (3.14159265*2.0f/4.0f)/256.0f;
-    ab += (3.14159265*2.0f/4.0f)/256.0f;
-  }
-  panning_left[255] = 0;
-  panning_right[0] = 0;
-}
-
-void hvl_GenSawtooth( int8 *buf, uint32 len )
-{
-  uint32 i;
-  int32  val, add;
-  
-  add = 256 / (len-1);
-  val = -128;
-  
-  for( i=0; i<len; i++, val += add )
-    *buf++ = (int8)val;  
-}
-
-void hvl_GenTriangle( int8 *buf, uint32 len )
-{
-  uint32 i;
-  int32  d2, d5, d1, d4;
-  int32  val;
-  int8   *buf2;
-  
-  d2  = len;
-  d5  = len >> 2;
-  d1  = 128/d5;
-  d4  = -(d2 >> 1);
-  val = 0;
-  
-  for( i=0; i<d5; i++ )
-  {
-    *buf++ = val;
-    val += d1;
-  }
-  *buf++ = 0x7f;
-
-  if( d5 != 1 )
-  {
-    val = 128;
-    for( i=0; i<d5-1; i++ )
-    {
-      val -= d1;
-      *buf++ = val;
-    }
-  }
-  
-  buf2 = buf + d4;
-  for( i=0; i<d5*2; i++ )
-  {
-    int8 c;
-    
-    c = *buf2++;
-    if( c == 0x7f )
-      c = 0x80;
-    else
-      c = -c;
-    
-    *buf++ = c;
-  }
-}
-
-void hvl_GenSquare( int8 *buf )
-{
-  uint32 i, j;
-  
-  for( i=1; i<=0x20; i++ )
-  {
-    for( j=0; j<(0x40-i)*2; j++ )
-      *buf++ = 0x80;
-    for( j=0; j<i*2; j++ )
-      *buf++ = 0x7f;
-  }
-}
-
-static inline int32 clipshifted8(int32 in)
-{
-  int16 top = (int16)(in >> 16);
-  if (top > 127) in = 127 << 16;
-  else if (top < -128) in = -(128 << 16);
-  return in;
-}
-
-void hvl_GenFilterWaves( int8 *buf, int8 *lowbuf, int8 *highbuf )
-{
-
-
-  const int16 * mid_table = &filter_thing[0];
-  const int16 * low_table = &filter_thing[1395];
-
-  int32 freq;
-  int32 i;
-
-  for( i=0, freq = 25; i<31; i++, freq += 9 )
-  {
-    uint32 wv;
-    int8  *a0 = buf;
-
-    for( wv=0; wv<6+6+0x20+1; wv++ )
-    {
-      int32 in, fre, high, mid, low;
-      uint32  j;
-
-      mid  = *mid_table++ << 8;
-      low = *low_table++ << 8;
-
-      for( j=0; j<=lentab[wv]; j++ )
-      {
-        in   = a0[j] << 16;
-        high = clipshifted8( in - mid - low );
-        fre  = (high >> 8) * freq;
-        mid  = clipshifted8(mid + fre);
-        fre  = (mid  >> 8) * freq;
-        low  = clipshifted8(low + fre);
-        *highbuf++ = high >> 16;
-        *lowbuf++  = low  >> 16;
-      }
-      a0 += lentab[wv]+1;
-    }
-  }
-}
-
-void hvl_GenWhiteNoise( int8 *buf, uint32 len )
-{
-  uint32 ays;
-
-  ays = 0x41595321;
-
-  do {
-    uint16 ax, bx;
-    int8 s;
-
-    s = ays;
-
-    if( ays & 0x100 )
-    {
-      s = 0x7f;
-
-      if( ays & 0x8000 )
-        s = 0x80;
-    }
-
-    *buf++ = s;
-    len--;
-
-    ays = (ays >> 5) | (ays << 27);
-    ays = (ays & 0xffffff00) | ((ays & 0xff) ^ 0x9a);
-    bx  = ays;
-    ays = (ays << 2) | (ays >> 30);
-    ax  = ays;
-    bx  += ax;
-    ax  ^= bx;
-    ays  = (ays & 0xffff0000) | ax;
-    ays  = (ays >> 3) | (ays << 29);
-  } while( len );
-}
 
 void hvl_reset_some_stuff( struct hvl_tune *ht )
 {
@@ -238,7 +59,7 @@ void hvl_reset_some_stuff( struct hvl_tune *ht )
     memset(&ht->ht_Voices[i].vc_VoiceBuffer,0,0x281);
     memset(&ht->ht_Voices[i].vc_RingVoiceBuffer,0,0x281);
   }
-  
+
   for( i=0; i<MAX_CHANNELS; i++ )
   {
     ht->ht_Voices[i].vc_WNRandom          = 0x280;
@@ -257,10 +78,10 @@ BOOL hvl_InitSubsong( struct hvl_tune *ht, uint32 nr )
     return FALSE;
 
   ht->ht_SongNum = nr;
-  
+
   PosNr = 0;
   if( nr ) PosNr = ht->ht_Subsongs[nr-1];
-  
+
   ht->ht_PosNr          = PosNr;
   ht->ht_PosJump        = 0;
   ht->ht_PatternBreak   = 0;
@@ -271,7 +92,7 @@ BOOL hvl_InitSubsong( struct hvl_tune *ht, uint32 nr )
   ht->ht_GetNewPosition = 1;
   ht->ht_SongEndReached = 0;
   ht->ht_PlayingTime    = 0;
-  
+
   for( i=0; i<MAX_CHANNELS; i+=4 )
   {
     ht->ht_Voices[i+0].vc_Pan          = ht->ht_defpanleft;
@@ -293,28 +114,13 @@ BOOL hvl_InitSubsong( struct hvl_tune *ht, uint32 nr )
   }
 
   hvl_reset_some_stuff( ht );
-  
+
   return TRUE;
 }
 
 void hvl_InitReplayer( void )
 {
-  hvl_GenPanningTables();
-  hvl_GenSawtooth( &waves[WO_SAWTOOTH_04], 0x04 );
-  hvl_GenSawtooth( &waves[WO_SAWTOOTH_08], 0x08 );
-  hvl_GenSawtooth( &waves[WO_SAWTOOTH_10], 0x10 );
-  hvl_GenSawtooth( &waves[WO_SAWTOOTH_20], 0x20 );
-  hvl_GenSawtooth( &waves[WO_SAWTOOTH_40], 0x40 );
-  hvl_GenSawtooth( &waves[WO_SAWTOOTH_80], 0x80 );
-  hvl_GenTriangle( &waves[WO_TRIANGLE_04], 0x04 );
-  hvl_GenTriangle( &waves[WO_TRIANGLE_08], 0x08 );
-  hvl_GenTriangle( &waves[WO_TRIANGLE_10], 0x10 );
-  hvl_GenTriangle( &waves[WO_TRIANGLE_20], 0x20 );
-  hvl_GenTriangle( &waves[WO_TRIANGLE_40], 0x40 );
-  hvl_GenTriangle( &waves[WO_TRIANGLE_80], 0x80 );
-  hvl_GenSquare( &waves[WO_SQUARES] );
-  hvl_GenWhiteNoise( &waves[WO_WHITENOISE], WHITENOISELEN );
-  hvl_GenFilterWaves( &waves[WO_TRIANGLE_04], &waves[WO_LOWPASSES], &waves[WO_HIGHPASSES] );
+  hvl_GenTables();
 }
 
 struct hvl_tune *hvl_load_ahx( uint8 *buf, uint32 buflen, uint32 defstereo, uint32 freq )
@@ -325,7 +131,7 @@ struct hvl_tune *hvl_load_ahx( uint8 *buf, uint32 buflen, uint32 defstereo, uint
   struct hvl_tune *ht;
   struct  hvl_plsentry *ple;
   int32 defgain[] = { 71, 72, 76, 85, 100 };
-  
+
   posn = ((buf[6]&0x0f)<<8)|buf[7];
   insn = buf[12];
   ssn  = buf[13];
@@ -343,7 +149,7 @@ struct hvl_tune *hvl_load_ahx( uint8 *buf, uint32 buflen, uint32 defstereo, uint
   bptr += posn*4*2; // Skip past the positions
   bptr += trkn*trkl*3;
   if((buf[6]&0x80)==0) bptr += trkl*3;
-  
+
   // *NOW* we can finally calculate PList space
   for( i=1; i<=insn; i++ )
   {
@@ -358,10 +164,10 @@ struct hvl_tune *hvl_load_ahx( uint8 *buf, uint32 buflen, uint32 defstereo, uint
     printf( "Out of memory!\n" );
     return NULL;
   }
-  
+
   ht->ht_Frequency       = freq;
   ht->ht_FreqF           = (float64)freq;
-  
+
   ht->ht_Positions   = (struct hvl_position *)(&ht[1]);
   ht->ht_Instruments = (struct hvl_instrument *)(&ht->ht_Positions[posn]);
   ht->ht_Subsongs    = (uint16 *)(&ht->ht_Instruments[(insn+1)]);
@@ -384,11 +190,11 @@ struct hvl_tune *hvl_load_ahx( uint8 *buf, uint32 buflen, uint32 defstereo, uint
   ht->ht_defpanleft      = stereopan_left[ht->ht_defstereo];
   ht->ht_defpanright     = stereopan_right[ht->ht_defstereo];
   ht->ht_mixgain         = (defgain[ht->ht_defstereo]*256)/100;
-  
+
   if( ht->ht_Restart >= ht->ht_PositionNr )
     ht->ht_Restart = ht->ht_PositionNr-1;
 
-  // Do some validation  
+  // Do some validation
   if( ( ht->ht_PositionNr > 1000 ) ||
       ( ht->ht_TrackLength > 64 ) ||
       ( ht->ht_InstrumentNr > 64 ) )
@@ -406,7 +212,7 @@ struct hvl_tune *hvl_load_ahx( uint8 *buf, uint32 buflen, uint32 defstereo, uint
   nptr = (TEXT *)&buf[((buf[4]<<8)|buf[5])+strlen( ht->ht_Name )+1];
 
   bptr = &buf[14];
-  
+
   // Subsongs
   for( i=0; i<ht->ht_SubsongNr; i++ )
   {
@@ -415,7 +221,7 @@ struct hvl_tune *hvl_load_ahx( uint8 *buf, uint32 buflen, uint32 defstereo, uint
       ht->ht_Subsongs[i] = 0;
     bptr += 2;
   }
-  
+
   // Position list
   for( i=0; i<ht->ht_PositionNr; i++ )
   {
@@ -425,7 +231,7 @@ struct hvl_tune *hvl_load_ahx( uint8 *buf, uint32 buflen, uint32 defstereo, uint
       ht->ht_Positions[i].pos_Transpose[j] = *(int8 *)bptr++;
     }
   }
-  
+
   // Tracks
   for( i=0; i<=ht->ht_TrackNr; i++ )
   {
@@ -442,7 +248,7 @@ struct hvl_tune *hvl_load_ahx( uint8 *buf, uint32 buflen, uint32 defstereo, uint
       }
       continue;
     }
-    
+
     for( j=0; j<ht->ht_TrackLength; j++ )
     {
       ht->ht_Tracks[i][j].stp_Note       = (bptr[0]>>2)&0x3f;
@@ -454,7 +260,7 @@ struct hvl_tune *hvl_load_ahx( uint8 *buf, uint32 buflen, uint32 defstereo, uint
       bptr += 3;
     }
   }
-  
+
   // Instruments
   for( i=1; i<=ht->ht_InstrumentNr; i++ )
   {
@@ -465,7 +271,7 @@ struct hvl_tune *hvl_load_ahx( uint8 *buf, uint32 buflen, uint32 defstereo, uint
     } else {
       ht->ht_Instruments[i].ins_Name[0] = 0;
     }
-    
+
     ht->ht_Instruments[i].ins_Volume      = bptr[0];
     ht->ht_Instruments[i].ins_FilterSpeed = ((bptr[1]>>3)&0x1f)|((bptr[12]>>2)&0x20);
     ht->ht_Instruments[i].ins_WaveLength  = bptr[1]&0x07;
@@ -477,7 +283,7 @@ struct hvl_tune *hvl_load_ahx( uint8 *buf, uint32 buflen, uint32 defstereo, uint
     ht->ht_Instruments[i].ins_Envelope.sFrames = bptr[6];
     ht->ht_Instruments[i].ins_Envelope.rFrames = bptr[7];
     ht->ht_Instruments[i].ins_Envelope.rVolume = bptr[8];
-    
+
     ht->ht_Instruments[i].ins_FilterLowerLimit     = bptr[12]&0x7f;
     ht->ht_Instruments[i].ins_VibratoDelay         = bptr[13];
     ht->ht_Instruments[i].ins_HardCutReleaseFrames = (bptr[14]>>4)&0x07;
@@ -490,10 +296,10 @@ struct hvl_tune *hvl_load_ahx( uint8 *buf, uint32 buflen, uint32 defstereo, uint
     ht->ht_Instruments[i].ins_FilterUpperLimit     = bptr[19]&0x3f;
     ht->ht_Instruments[i].ins_PList.pls_Speed      = bptr[20];
     ht->ht_Instruments[i].ins_PList.pls_Length     = bptr[21];
-    
+
     ht->ht_Instruments[i].ins_PList.pls_Entries    = ple;
     ple += bptr[21];
-    
+
     bptr += 22;
     for( j=0; j<ht->ht_Instruments[i].ins_PList.pls_Length; j++ )
     {
@@ -521,7 +327,7 @@ struct hvl_tune *hvl_load_ahx( uint8 *buf, uint32 buflen, uint32 defstereo, uint
       bptr += 4;
     }
   }
-  
+
   hvl_InitSubsong( ht, 0 );
   IExec->FreeVec( buf );
   return ht;
@@ -543,7 +349,7 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
     printf( "Out of memory\n" );
     return NULL;
   }
-  
+
   lock = IDOS->Lock( name, ACCESS_READ );
   if( !lock )
   {
@@ -551,7 +357,7 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
     printf( "Unable to find '%s'\n", name );
     return NULL;
   }
-  
+
   IDOS->Examine( lock, fib );
   if( !FIB_IS_FILE( fib ) )
   {
@@ -560,7 +366,7 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
     printf( "Bad file!\n" );
     return NULL;
   }
-  
+
   buflen = fib->fib_Size;
   buf = IExec->AllocVec( buflen, MEMF_ANY );
   if( !buf )
@@ -570,7 +376,7 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
     printf( "Out of memory!\n" );
     return NULL;
   }
-  
+
   fh = IDOS->OpenFromLock( lock );
   if( !fh )
   {
@@ -591,7 +397,7 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
   }
   IDOS->Close( fh );
   IDOS->FreeDosObject( DOS_FIB, fib );
-  
+
   if( ( buf[0] == 'T' ) &&
       ( buf[1] == 'H' ) &&
       ( buf[2] == 'X' ) &&
@@ -607,7 +413,7 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
     printf( "Invalid file.\n" );
     return NULL;
   }
-  
+
   posn = ((buf[6]&0x0f)<<8)|buf[7];
   insn = buf[12];
   ssn  = buf[13];
@@ -624,7 +430,7 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
   bptr = &buf[16];
   bptr += ssn*2;       // Skip past the subsong list
   bptr += posn*chnn*2; // Skip past the positions
-  
+
   // Skip past the tracks
   // 1.4: Fixed two really stupid bugs that cancelled each other
   //      out if the module had a blank first track (which is how
@@ -639,26 +445,26 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
       }
       bptr += 5;
     }
-  
+
   // *NOW* we can finally calculate PList space
   for( i=1; i<=insn; i++ )
   {
     hs += bptr[21] * sizeof( struct hvl_plsentry );
     bptr += 22 + bptr[21]*5;
   }
-  
-  ht = IExec->AllocVec( hs, MEMF_ANY );    
+
+  ht = IExec->AllocVec( hs, MEMF_ANY );
   if( !ht )
   {
     IExec->FreeVec( buf );
     printf( "Out of memory!\n" );
     return NULL;
   }
-  
+
   ht->ht_Version         = buf[3]; // 1.5
   ht->ht_Frequency       = freq;
   ht->ht_FreqF           = (float64)freq;
-  
+
   ht->ht_Positions       = (struct hvl_position *)(&ht[1]);
   ht->ht_Instruments     = (struct hvl_instrument *)(&ht->ht_Positions[posn]);
   ht->ht_Subsongs        = (uint16 *)(&ht->ht_Instruments[(insn+1)]);
@@ -680,11 +486,11 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
   ht->ht_defstereo       = buf[15];
   ht->ht_defpanleft      = stereopan_left[ht->ht_defstereo];
   ht->ht_defpanright     = stereopan_right[ht->ht_defstereo];
-  
+
   if( ht->ht_Restart >= ht->ht_PositionNr )
     ht->ht_Restart = ht->ht_PositionNr-1;
 
-  // Do some validation  
+  // Do some validation
   if( ( ht->ht_PositionNr > 1000 ) ||
       ( ht->ht_TrackLength > 64 ) ||
       ( ht->ht_InstrumentNr > 64 ) )
@@ -703,14 +509,14 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
 
 
   bptr = &buf[16];
-  
+
   // Subsongs
   for( i=0; i<ht->ht_SubsongNr; i++ )
   {
     ht->ht_Subsongs[i] = (bptr[0]<<8)|bptr[1];
     bptr += 2;
   }
-  
+
   // Position list
   for( i=0; i<ht->ht_PositionNr; i++ )
   {
@@ -720,7 +526,7 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
       ht->ht_Positions[i].pos_Transpose[j] = *(int8 *)bptr++;
     }
   }
-  
+
   // Tracks
   for( i=0; i<=ht->ht_TrackNr; i++ )
   {
@@ -737,7 +543,7 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
       }
       continue;
     }
-    
+
     for( j=0; j<ht->ht_TrackLength; j++ )
     {
       if( bptr[0] == 0x3f )
@@ -751,7 +557,7 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
         bptr++;
         continue;
       }
-      
+
       ht->ht_Tracks[i][j].stp_Note       = bptr[0];
       ht->ht_Tracks[i][j].stp_Instrument = bptr[1];
       ht->ht_Tracks[i][j].stp_FX         = bptr[2]>>4;
@@ -761,8 +567,8 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
       bptr += 5;
     }
   }
-  
-  
+
+
   // Instruments
   for( i=1; i<=ht->ht_InstrumentNr; i++ )
   {
@@ -773,7 +579,7 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
     } else {
       ht->ht_Instruments[i].ins_Name[0] = 0;
     }
-    
+
     ht->ht_Instruments[i].ins_Volume      = bptr[0];
     ht->ht_Instruments[i].ins_FilterSpeed = ((bptr[1]>>3)&0x1f)|((bptr[12]>>2)&0x20);
     ht->ht_Instruments[i].ins_WaveLength  = bptr[1]&0x07;
@@ -785,7 +591,7 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
     ht->ht_Instruments[i].ins_Envelope.sFrames = bptr[6];
     ht->ht_Instruments[i].ins_Envelope.rFrames = bptr[7];
     ht->ht_Instruments[i].ins_Envelope.rVolume = bptr[8];
-    
+
     ht->ht_Instruments[i].ins_FilterLowerLimit     = bptr[12]&0x7f;
     ht->ht_Instruments[i].ins_VibratoDelay         = bptr[13];
     ht->ht_Instruments[i].ins_HardCutReleaseFrames = (bptr[14]>>4)&0x07;
@@ -798,10 +604,10 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
     ht->ht_Instruments[i].ins_FilterUpperLimit     = bptr[19]&0x3f;
     ht->ht_Instruments[i].ins_PList.pls_Speed      = bptr[20];
     ht->ht_Instruments[i].ins_PList.pls_Length     = bptr[21];
-    
+
     ht->ht_Instruments[i].ins_PList.pls_Entries    = ple;
     ple += bptr[21];
-    
+
     bptr += 22;
     for( j=0; j<ht->ht_Instruments[i].ins_PList.pls_Length; j++ )
     {
@@ -815,7 +621,7 @@ struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo )
       bptr += 5;
     }
   }
-  
+
   hvl_InitSubsong( ht, 0 );
   IExec->FreeVec( buf );
   return ht;
@@ -841,7 +647,7 @@ void hvl_process_stepfx_1( struct hvl_tune *ht, struct hvl_voice *voice, int32 F
       voice->vc_VolumeSlideDown = FXParam & 0x0f;
       voice->vc_VolumeSlideUp   = FXParam >> 4;
       break;
-    
+
     case 0x7:  // Panning
       if( FXParam > 127 )
         FXParam -= 256;
@@ -850,14 +656,14 @@ void hvl_process_stepfx_1( struct hvl_tune *ht, struct hvl_voice *voice, int32 F
       voice->vc_PanMultLeft  = panning_left[voice->vc_Pan];
       voice->vc_PanMultRight = panning_right[voice->vc_Pan];
       break;
-    
+
     case 0xb: // Position jump
       ht->ht_PosJump      = ht->ht_PosJump*100 + (FXParam & 0x0f) + (FXParam >> 4)*10;
       ht->ht_PatternBreak = 1;
       if( ht->ht_PosJump <= ht->ht_PosNr )
         ht->ht_SongEndReached = 1;
       break;
-    
+
     case 0xd: // Pattern break
       ht->ht_PosJump      = ht->ht_PosNr+1;
       ht->ht_PosJumpNote  = (FXParam & 0x0f) + (FXParam>>4)*10;
@@ -865,7 +671,7 @@ void hvl_process_stepfx_1( struct hvl_tune *ht, struct hvl_voice *voice, int32 F
       if( ht->ht_PosJumpNote >  ht->ht_TrackLength )
         ht->ht_PosJumpNote = 0;
       break;
-    
+
     case 0xe: // Extended commands
       switch( FXParam >> 4 )
       {
@@ -880,17 +686,17 @@ void hvl_process_stepfx_1( struct hvl_tune *ht, struct hvl_voice *voice, int32 F
             }
           }
           break;
-          
+
           // 1.6: 0xd case removed
       }
       break;
-    
+
     case 0xf: // Speed
       ht->ht_Tempo = FXParam;
       if( FXParam == 0 )
         ht->ht_SongEndReached = 1;
       break;
-  }  
+  }
 }
 
 void hvl_process_stepfx_2( struct hvl_tune *ht, struct hvl_voice *voice, int32 FX, int32 FXParam, int32 *Note )
@@ -902,11 +708,11 @@ void hvl_process_stepfx_2( struct hvl_tune *ht, struct hvl_voice *voice, int32 F
 //      voice->vc_PlantSquare  = 1;
       voice->vc_IgnoreSquare = 1;
       break;
-    
+
     case 0x3: // Tone portamento
       if( FXParam != 0 ) voice->vc_PeriodSlideSpeed = FXParam;
     case 0x5: // Tone portamento + volume slide
-      
+
       if( *Note )
       {
         int32 new, diff;
@@ -915,21 +721,21 @@ void hvl_process_stepfx_2( struct hvl_tune *ht, struct hvl_voice *voice, int32 F
         diff  = period_tab[voice->vc_TrackPeriod];
         diff -= new;
         new   = diff + voice->vc_PeriodSlidePeriod;
-        
+
         if( new )
           voice->vc_PeriodSlideLimit = -diff;
       }
       voice->vc_PeriodSlideOn        = 1;
       voice->vc_PeriodSlideWithLimit = 1;
       *Note = 0;
-      break;      
-  } 
+      break;
+  }
 }
 
 void hvl_process_stepfx_3( struct hvl_tune *ht, struct hvl_voice *voice, int32 FX, int32 FXParam )
 {
   int32 i;
-  
+
   switch( FX )
   {
     case 0x01: // Portamento up (period slide down)
@@ -959,7 +765,7 @@ void hvl_process_stepfx_3( struct hvl_tune *ht, struct hvl_voice *voice, int32 F
         voice->vc_NoteMaxVolume = FXParam;
         break;
       }
-      
+
       if( (FXParam -= 0x50) < 0 ) break;  // 1.6
 
       if( FXParam <= 0x40 )
@@ -968,7 +774,7 @@ void hvl_process_stepfx_3( struct hvl_tune *ht, struct hvl_voice *voice, int32 F
           ht->ht_Voices[i].vc_TrackMasterVolume = FXParam;
         break;
       }
-      
+
       if( (FXParam -= 0xa0-0x50) < 0 ) break; // 1.6
 
       if( FXParam <= 0x40 )
@@ -982,30 +788,30 @@ void hvl_process_stepfx_3( struct hvl_tune *ht, struct hvl_voice *voice, int32 F
           voice->vc_PeriodSlidePeriod -= (FXParam & 0x0f); // 1.8
           voice->vc_PlantPeriod = 1;
           break;
-        
+
         case 0x2: // Fineslide down
           voice->vc_PeriodSlidePeriod += (FXParam & 0x0f); // 1.8
           voice->vc_PlantPeriod = 1;
           break;
-        
+
         case 0x4: // Vibrato control
           voice->vc_VibratoDepth = FXParam & 0x0f;
           break;
-        
+
         case 0x0a: // Fine volume up
           voice->vc_NoteMaxVolume += FXParam & 0x0f;
-          
+
           if( voice->vc_NoteMaxVolume > 0x40 )
             voice->vc_NoteMaxVolume = 0x40;
           break;
-        
+
         case 0x0b: // Fine volume down
           voice->vc_NoteMaxVolume -= FXParam & 0x0f;
-          
+
           if( voice->vc_NoteMaxVolume < 0 )
             voice->vc_NoteMaxVolume = 0;
           break;
-        
+
         case 0x0f: // Misc flags (1.5)
           if( ht->ht_Version < 1 ) break;
           switch( FXParam & 0xf )
@@ -1015,7 +821,7 @@ void hvl_process_stepfx_3( struct hvl_tune *ht, struct hvl_voice *voice, int32 F
               break;
           }
           break;
-      } 
+      }
       break;
   }
 }
@@ -1024,17 +830,17 @@ void hvl_process_step( struct hvl_tune *ht, struct hvl_voice *voice )
 {
   int32  Note, Instr, donenotedel;
   struct hvl_step *Step;
-  
+
   if( voice->vc_TrackOn == 0 )
     return;
-  
+
   voice->vc_VolumeSlideUp = voice->vc_VolumeSlideDown = 0;
-  
+
   Step = &ht->ht_Tracks[ht->ht_Positions[ht->ht_PosNr].pos_Track[voice->vc_VoiceNum]][ht->ht_NoteNr];
-  
+
   Note    = Step->stp_Note;
   Instr   = Step->stp_Instrument;
-  
+
   // --------- 1.6: from here --------------
 
   donenotedel = 0;
@@ -1081,14 +887,14 @@ void hvl_process_step( struct hvl_tune *ht, struct hvl_voice *voice )
 
   if( Note ) voice->vc_OverrideTranspose = 1000; // 1.5
 
-  hvl_process_stepfx_1( ht, voice, Step->stp_FX&0xf,  Step->stp_FXParam );  
+  hvl_process_stepfx_1( ht, voice, Step->stp_FX&0xf,  Step->stp_FXParam );
   hvl_process_stepfx_1( ht, voice, Step->stp_FXb&0xf, Step->stp_FXbParam );
-  
+
   if( ( Instr ) && ( Instr <= ht->ht_InstrumentNr ) )
   {
     struct hvl_instrument *Ins;
     int16  SquareLower, SquareUpper, d6, d3, d4;
-    
+
     /* 1.4: Reset panning to last set position */
     voice->vc_Pan          = voice->vc_SetPan;
     voice->vc_PanMultLeft  = panning_left[voice->vc_Pan];
@@ -1100,7 +906,7 @@ void hvl_process_step( struct hvl_tune *ht, struct hvl_voice *voice )
     voice->vc_ADSRVolume       = 0;
     voice->vc_Instrument       = Ins = &ht->ht_Instruments[Instr];
     voice->vc_SamplePos        = 0;
-    
+
     voice->vc_ADSR.aFrames     = Ins->ins_Envelope.aFrames;
 	voice->vc_ADSR.aVolume     = voice->vc_ADSR.aFrames ? Ins->ins_Envelope.aVolume*256/voice->vc_ADSR.aFrames : Ins->ins_Envelope.aVolume * 256; // XXX
     voice->vc_ADSR.dFrames     = Ins->ins_Envelope.dFrames;
@@ -1108,73 +914,73 @@ void hvl_process_step( struct hvl_tune *ht, struct hvl_voice *voice )
     voice->vc_ADSR.sFrames     = Ins->ins_Envelope.sFrames;
     voice->vc_ADSR.rFrames     = Ins->ins_Envelope.rFrames;
 	voice->vc_ADSR.rVolume     = voice->vc_ADSR.rFrames ? (Ins->ins_Envelope.rVolume-Ins->ins_Envelope.dVolume)*256/voice->vc_ADSR.rFrames : Ins->ins_Envelope.rVolume * 256; // XXX
-    
+
     voice->vc_WaveLength       = Ins->ins_WaveLength;
     voice->vc_NoteMaxVolume    = Ins->ins_Volume;
-    
+
     voice->vc_VibratoCurrent   = 0;
     voice->vc_VibratoDelay     = Ins->ins_VibratoDelay;
     voice->vc_VibratoDepth     = Ins->ins_VibratoDepth;
     voice->vc_VibratoSpeed     = Ins->ins_VibratoSpeed;
     voice->vc_VibratoPeriod    = 0;
-    
+
     voice->vc_HardCutRelease   = Ins->ins_HardCutRelease;
     voice->vc_HardCut          = Ins->ins_HardCutReleaseFrames;
-    
+
     voice->vc_IgnoreSquare = voice->vc_SquareSlidingIn = 0;
     voice->vc_SquareWait   = voice->vc_SquareOn        = 0;
-    
+
     SquareLower = Ins->ins_SquareLowerLimit >> (5 - voice->vc_WaveLength);
     SquareUpper = Ins->ins_SquareUpperLimit >> (5 - voice->vc_WaveLength);
-    
+
     if( SquareUpper < SquareLower )
     {
       int16 t = SquareUpper;
       SquareUpper = SquareLower;
       SquareLower = t;
     }
-    
+
     voice->vc_SquareUpperLimit = SquareUpper;
     voice->vc_SquareLowerLimit = SquareLower;
-    
+
     voice->vc_IgnoreFilter    = voice->vc_FilterWait = voice->vc_FilterOn = 0;
     voice->vc_FilterSlidingIn = 0;
 
     d6 = Ins->ins_FilterSpeed;
     d3 = Ins->ins_FilterLowerLimit;
     d4 = Ins->ins_FilterUpperLimit;
-    
+
     if( d3 & 0x80 ) d6 |= 0x20;
     if( d4 & 0x80 ) d6 |= 0x40;
-    
+
     voice->vc_FilterSpeed = d6;
     d3 &= ~0x80;
     d4 &= ~0x80;
-    
+
     if( d3 > d4 )
     {
       int16 t = d3;
       d3 = d4;
       d4 = t;
     }
-    
+
     voice->vc_FilterUpperLimit = d4;
     voice->vc_FilterLowerLimit = d3;
     voice->vc_FilterPos        = 32;
-    
+
     voice->vc_PerfWait  = voice->vc_PerfCurrent = 0;
     voice->vc_PerfSpeed = Ins->ins_PList.pls_Speed;
     voice->vc_PerfList  = &voice->vc_Instrument->ins_PList;
-    
+
     voice->vc_RingMixSource   = NULL;   // No ring modulation
     voice->vc_RingSamplePos   = 0;
     voice->vc_RingPlantPeriod = 0;
     voice->vc_RingNewWaveform = 0;
   }
-  
+
   voice->vc_PeriodSlideOn = 0;
-  
-  hvl_process_stepfx_2( ht, voice, Step->stp_FX&0xf,  Step->stp_FXParam,  &Note );  
+
+  hvl_process_stepfx_2( ht, voice, Step->stp_FX&0xf,  Step->stp_FXParam,  &Note );
   hvl_process_stepfx_2( ht, voice, Step->stp_FXb&0xf, Step->stp_FXbParam, &Note );
 
   if( Note )
@@ -1182,9 +988,9 @@ void hvl_process_step( struct hvl_tune *ht, struct hvl_voice *voice )
     voice->vc_TrackPeriod = Note;
     voice->vc_PlantPeriod = 1;
   }
-  
-  hvl_process_stepfx_3( ht, voice, Step->stp_FX&0xf,  Step->stp_FXParam );  
-  hvl_process_stepfx_3( ht, voice, Step->stp_FXb&0xf, Step->stp_FXbParam );  
+
+  hvl_process_stepfx_3( ht, voice, Step->stp_FX&0xf,  Step->stp_FXParam );
+  hvl_process_stepfx_3( ht, voice, Step->stp_FXb&0xf, Step->stp_FXbParam );
 }
 
 void hvl_plist_command_parse( struct hvl_tune *ht, struct hvl_voice *voice, int32 FX, int32 FXParam )
@@ -1214,14 +1020,14 @@ void hvl_plist_command_parse( struct hvl_tune *ht, struct hvl_voice *voice, int3
       voice->vc_PeriodPerfSlideSpeed = -FXParam;
       voice->vc_PeriodPerfSlideOn    = 1;
       break;
-    
+
     case 3:
       if( voice->vc_IgnoreSquare == 0 )
         voice->vc_SquarePos = FXParam >> (5-voice->vc_WaveLength);
       else
         voice->vc_IgnoreSquare = 0;
       break;
-    
+
     case 4:
       if( FXParam == 0 )
       {
@@ -1236,7 +1042,7 @@ void hvl_plist_command_parse( struct hvl_tune *ht, struct hvl_voice *voice, int3
           if(( FXParam & 0x0f ) == 0x0f )
             voice->vc_SquareSign = -1;
         }
-        
+
         if( FXParam & 0xf0 )
         {
           voice->vc_FilterInit = (voice->vc_FilterOn ^= 1);
@@ -1246,11 +1052,11 @@ void hvl_plist_command_parse( struct hvl_tune *ht, struct hvl_voice *voice, int3
         }
       }
       break;
-    
+
     case 5:
       voice->vc_PerfCurrent = FXParam;
       break;
-    
+
     case 7:
       // Ring modulate with triangle
       if(( FXParam >= 1 ) && ( FXParam <= 0x3C ))
@@ -1267,12 +1073,12 @@ void hvl_plist_command_parse( struct hvl_tune *ht, struct hvl_voice *voice, int3
         voice->vc_RingAudioSource = NULL; // turn it off
         voice->vc_RingMixSource   = NULL;
         break;
-      }    
+      }
       voice->vc_RingWaveform    = 0;
       voice->vc_RingNewWaveform = 1;
       voice->vc_RingPlantPeriod = 1;
       break;
-    
+
     case 8:  // Ring modulate with sawtooth
       if(( FXParam >= 1 ) && ( FXParam <= 0x3C ))
       {
@@ -1295,8 +1101,8 @@ void hvl_plist_command_parse( struct hvl_tune *ht, struct hvl_voice *voice, int3
       voice->vc_RingPlantPeriod = 1;
       break;
 
-    /* New in HivelyTracker 1.4 */    
-    case 9:    
+    /* New in HivelyTracker 1.4 */
+    case 9:
       if( FXParam > 127 )
         FXParam -= 256;
       voice->vc_Pan          = (FXParam+128);
@@ -1310,7 +1116,7 @@ void hvl_plist_command_parse( struct hvl_tune *ht, struct hvl_voice *voice, int3
         voice->vc_NoteMaxVolume = FXParam;
         break;
       }
-      
+
       if( (FXParam -= 0x50) < 0 ) break;
 
       if( FXParam <= 0x40 )
@@ -1318,17 +1124,17 @@ void hvl_plist_command_parse( struct hvl_tune *ht, struct hvl_voice *voice, int3
         voice->vc_PerfSubVolume = FXParam;
         break;
       }
-      
+
       if( (FXParam -= 0xa0-0x50) < 0 ) break;
-      
+
       if( FXParam <= 0x40 )
         voice->vc_TrackMasterVolume = FXParam;
       break;
-    
+
     case 15:
       voice->vc_PerfSpeed = voice->vc_PerfWait = FXParam;
       break;
-  } 
+  }
 }
 
 void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
@@ -1345,24 +1151,24 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
     else
       voice->vc_NoteDelayWait--;
   }
-  
+
   if( voice->vc_HardCut )
   {
     int32 nextinst;
-    
+
     if( ht->ht_NoteNr+1 < ht->ht_TrackLength )
       nextinst = ht->ht_Tracks[voice->vc_Track][ht->ht_NoteNr+1].stp_Instrument;
     else
       nextinst = ht->ht_Tracks[voice->vc_NextTrack][0].stp_Instrument;
-    
+
     if( nextinst )
     {
       int32 d1;
-      
+
       d1 = ht->ht_Tempo - voice->vc_HardCut;
-      
+
       if( d1 < 0 ) d1 = 0;
-    
+
       if( !voice->vc_NoteCutOn )
       {
         voice->vc_NoteCutOn       = 1;
@@ -1373,13 +1179,13 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
       }
     }
   }
-    
+
   if( voice->vc_NoteCutOn )
   {
     if( voice->vc_NoteCutWait <= 0 )
     {
       voice->vc_NoteCutOn = 0;
-        
+
       if( voice->vc_HardCutRelease )
       {
         voice->vc_ADSR.rVolume = -(voice->vc_ADSRVolume - (voice->vc_Instrument->ins_Envelope.rVolume << 8)) / voice->vc_HardCutReleaseF;
@@ -1392,30 +1198,30 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
       voice->vc_NoteCutWait--;
     }
   }
-    
+
   // ADSR envelope
   if( voice->vc_ADSR.aFrames )
   {
     voice->vc_ADSRVolume += voice->vc_ADSR.aVolume;
-      
+
     if( --voice->vc_ADSR.aFrames <= 0 )
       voice->vc_ADSRVolume = voice->vc_Instrument->ins_Envelope.aVolume << 8;
 
   } else if( voice->vc_ADSR.dFrames ) {
-    
+
     voice->vc_ADSRVolume += voice->vc_ADSR.dVolume;
-      
+
     if( --voice->vc_ADSR.dFrames <= 0 )
       voice->vc_ADSRVolume = voice->vc_Instrument->ins_Envelope.dVolume << 8;
-    
+
   } else if( voice->vc_ADSR.sFrames ) {
-    
+
     voice->vc_ADSR.sFrames--;
-    
+
   } else if( voice->vc_ADSR.rFrames ) {
-    
+
     voice->vc_ADSRVolume += voice->vc_ADSR.rVolume;
-    
+
     if( --voice->vc_ADSR.rFrames <= 0 )
       voice->vc_ADSRVolume = voice->vc_Instrument->ins_Envelope.rVolume << 8;
   }
@@ -1434,24 +1240,24 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
     if( voice->vc_PeriodSlideWithLimit )
     {
       int32  d0, d2;
-      
+
       d0 = voice->vc_PeriodSlidePeriod - voice->vc_PeriodSlideLimit;
       d2 = voice->vc_PeriodSlideSpeed;
-      
+
       if( d0 > 0 )
         d2 = -d2;
-      
+
       if( d0 )
       {
         int32 d3;
-         
+
         d3 = (d0 + d2) ^ d0;
-        
+
         if( d3 >= 0 )
           d0 = voice->vc_PeriodSlidePeriod + d2;
         else
           d0 = voice->vc_PeriodSlideLimit;
-        
+
         voice->vc_PeriodSlidePeriod = d0;
         voice->vc_PlantPeriod = 1;
       }
@@ -1460,7 +1266,7 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
       voice->vc_PlantPeriod = 1;
     }
   }
-  
+
   // Vibrato
   if( voice->vc_VibratoDepth )
   {
@@ -1473,7 +1279,7 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
       voice->vc_VibratoDelay--;
     }
   }
-  
+
   // PList
   if( voice->vc_PerfList != 0 )
   {
@@ -1483,23 +1289,23 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
       {
         uint32 i;
         int32 cur;
-        
+
         cur = voice->vc_PerfCurrent++;
         voice->vc_PerfWait = voice->vc_PerfSpeed;
-        
+
         if( voice->vc_PerfList->pls_Entries[cur].ple_Waveform )
         {
           voice->vc_Waveform             = voice->vc_PerfList->pls_Entries[cur].ple_Waveform-1;
           voice->vc_NewWaveform          = 1;
           voice->vc_PeriodPerfSlideSpeed = voice->vc_PeriodPerfSlidePeriod = 0;
         }
-        
+
         // Holdwave
         voice->vc_PeriodPerfSlideOn = 0;
-        
+
         for( i=0; i<2; i++ )
           hvl_plist_command_parse( ht, voice, voice->vc_PerfList->pls_Entries[cur].ple_FX[i]&0xff, voice->vc_PerfList->pls_Entries[cur].ple_FXParam[i]&0xff );
-        
+
         // GetNote
         if( voice->vc_PerfList->pls_Entries[cur].ple_Note )
         {
@@ -1515,30 +1321,30 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
         voice->vc_PeriodPerfSlideSpeed = 0;
     }
   }
-  
+
   // PerfPortamento
   if( voice->vc_PeriodPerfSlideOn )
   {
     voice->vc_PeriodPerfSlidePeriod -= voice->vc_PeriodPerfSlideSpeed;
-    
+
     if( voice->vc_PeriodPerfSlidePeriod )
       voice->vc_PlantPeriod = 1;
   }
-  
+
   if( voice->vc_Waveform == 3-1 && voice->vc_SquareOn )
   {
     if( --voice->vc_SquareWait <= 0 )
     {
       int32 d1, d2, d3;
-      
+
       d1 = voice->vc_SquareLowerLimit;
       d2 = voice->vc_SquareUpperLimit;
       d3 = voice->vc_SquarePos;
-      
+
       if( voice->vc_SquareInit )
       {
         voice->vc_SquareInit = 0;
-        
+
         if( d3 <= d1 )
         {
           voice->vc_SquareSlidingIn = 1;
@@ -1548,7 +1354,7 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
           voice->vc_SquareSign = -1;
         }
       }
-      
+
       // NoSquareInit
       if( d1 == d3 || d2 == d3 )
       {
@@ -1557,23 +1363,23 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
         else
           voice->vc_SquareSign = -voice->vc_SquareSign;
       }
-      
+
       d3 += voice->vc_SquareSign;
       voice->vc_SquarePos   = d3;
       voice->vc_PlantSquare = 1;
       voice->vc_SquareWait  = voice->vc_Instrument->ins_SquareSpeed;
     }
   }
-  
+
   if( voice->vc_FilterOn && --voice->vc_FilterWait <= 0 )
   {
     uint32 i, FMax;
     int32 d1, d2, d3;
-    
+
     d1 = voice->vc_FilterLowerLimit;
     d2 = voice->vc_FilterUpperLimit;
     d3 = voice->vc_FilterPos;
-    
+
     if( voice->vc_FilterInit )
     {
       voice->vc_FilterInit = 0;
@@ -1586,7 +1392,7 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
         voice->vc_FilterSign      = -1;
       }
     }
-    
+
     // NoFilterInit
     FMax = (voice->vc_FilterSpeed < 3) ? (5-voice->vc_FilterSpeed) : 1;
 
@@ -1601,13 +1407,13 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
       }
       d3 += voice->vc_FilterSign;
     }
-    
+
     if( d3 < 1 )  d3 = 1;
     if( d3 > 63 ) d3 = 63;
     voice->vc_FilterPos   = d3;
     voice->vc_NewWaveform = 1;
     voice->vc_FilterWait  = voice->vc_FilterSpeed - 3;
-    
+
     if( voice->vc_FilterWait < 1 )
       voice->vc_FilterWait = 1;
   }
@@ -1619,50 +1425,50 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
     int32   Delta;
     int8   *SquarePtr;
     int32  X;
-    
+
     SquarePtr = &waves[WO_SQUARES+(voice->vc_FilterPos-0x20)*(0xfc+0xfc+0x80*0x1f+0x80+0x280*3)];
     X = voice->vc_SquarePos << (5 - voice->vc_WaveLength);
-    
+
     if( X > 0x20 )
     {
       X = 0x40 - X;
       voice->vc_SquareReverse = 1;
     }
-    
+
     // OkDownSquare
     if( X > 0 )
       SquarePtr += (X-1) << 7;
-    
+
     Delta = 32 >> voice->vc_WaveLength;
     ht->ht_WaveformTab[2] = voice->vc_SquareTempBuffer;
-    
+
     for( i=0; i<(1<<voice->vc_WaveLength)*4; i++ )
     {
       voice->vc_SquareTempBuffer[i] = *SquarePtr;
       SquarePtr += Delta;
     }
-    
+
     voice->vc_NewWaveform = 1;
     voice->vc_Waveform    = 3-1;
     voice->vc_PlantSquare = 0;
   }
-  
+
   if( voice->vc_Waveform == 4-1 )
     voice->vc_NewWaveform = 1;
-  
+
   if( voice->vc_RingNewWaveform )
   {
     const int8 *rasrc;
-    
+
     if( voice->vc_RingWaveform > 1 ) voice->vc_RingWaveform = 1;
-    
+
     rasrc = ht->ht_WaveformTab[voice->vc_RingWaveform];
     rasrc += Offsets[voice->vc_WaveLength];
-    
+
     voice->vc_RingAudioSource = rasrc;
-  }    
-        
-  
+  }
+
+
   if( voice->vc_NewWaveform )
   {
     const int8 *AudioSource;
@@ -1689,12 +1495,12 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
 
     voice->vc_AudioSource = AudioSource;
   }
-  
+
   // Ring modulation period calculation
   if( voice->vc_RingAudioSource )
   {
     voice->vc_RingAudioPeriod = voice->vc_RingBasePeriod;
-  
+
     if( !(voice->vc_RingFixedPeriod) )
     {
       if( voice->vc_OverrideTranspose != 1000 )  // 1.5
@@ -1702,13 +1508,13 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
       else
         voice->vc_RingAudioPeriod += voice->vc_Transpose + voice->vc_TrackPeriod - 1;
     }
-  
+
     if( voice->vc_RingAudioPeriod > 5*12 )
       voice->vc_RingAudioPeriod = 5*12;
-  
+
     if( voice->vc_RingAudioPeriod < 0 )
       voice->vc_RingAudioPeriod = 0;
-  
+
     voice->vc_RingAudioPeriod = period_tab[voice->vc_RingAudioPeriod];
 
     if( !(voice->vc_RingFixedPeriod) )
@@ -1722,10 +1528,10 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
     if( voice->vc_RingAudioPeriod < 0x0071 )
       voice->vc_RingAudioPeriod = 0x0071;
   }
-  
+
   // Normal period calculation
   voice->vc_AudioPeriod = voice->vc_InstrPeriod;
-  
+
   if( !(voice->vc_FixedNote) )
   {
     if( voice->vc_OverrideTranspose != 1000 ) // 1.5
@@ -1733,26 +1539,26 @@ void hvl_process_frame( struct hvl_tune *ht, struct hvl_voice *voice )
     else
       voice->vc_AudioPeriod += voice->vc_Transpose + voice->vc_TrackPeriod - 1;
   }
-    
+
   if( voice->vc_AudioPeriod > 5*12 )
     voice->vc_AudioPeriod = 5*12;
-  
+
   if( voice->vc_AudioPeriod < 0 )
     voice->vc_AudioPeriod = 0;
-  
+
   voice->vc_AudioPeriod = period_tab[voice->vc_AudioPeriod];
-  
+
   if( !(voice->vc_FixedNote) )
     voice->vc_AudioPeriod += voice->vc_PeriodSlidePeriod;
 
-  voice->vc_AudioPeriod += voice->vc_PeriodPerfSlidePeriod + voice->vc_VibratoPeriod;    
+  voice->vc_AudioPeriod += voice->vc_PeriodPerfSlidePeriod + voice->vc_VibratoPeriod;
 
   if( voice->vc_AudioPeriod > 0x0d60 )
     voice->vc_AudioPeriod = 0x0d60;
 
   if( voice->vc_AudioPeriod < 0x0071 )
     voice->vc_AudioPeriod = 0x0071;
-  
+
   voice->vc_AudioVolume = (((((((voice->vc_ADSRVolume >> 8) * voice->vc_NoteMaxVolume) >> 6) * voice->vc_PerfSubVolume) >> 6) * voice->vc_TrackMasterVolume) >> 6);
 }
 
@@ -1763,17 +1569,17 @@ void hvl_set_audio( struct hvl_voice *voice, float64 freqf )
     voice->vc_VoiceVolume = 0;
     return;
   }
-  
+
   voice->vc_VoiceVolume = voice->vc_AudioVolume;
-  
+
   if( voice->vc_PlantPeriod )
   {
     float64 freq2;
     uint32  delta;
-    
+
     voice->vc_PlantPeriod = 0;
     voice->vc_VoicePeriod = voice->vc_AudioPeriod;
-    
+
     freq2 = Period2Freq( voice->vc_AudioPeriod );
     delta = (uint32)(freq2 / freqf);
 
@@ -1781,13 +1587,13 @@ void hvl_set_audio( struct hvl_voice *voice, float64 freqf )
     if( delta == 0 ) delta = 1;
     voice->vc_Delta = delta;
   }
-  
+
   if( voice->vc_NewWaveform )
   {
     const int8 *src;
-    
+
     src = voice->vc_AudioSource;
-    
+
     if( voice->vc_Waveform == 4-1 )
     {
       IExec->CopyMem((void *)src, &voice->vc_VoiceBuffer[0], 0x280);
@@ -1809,21 +1615,21 @@ void hvl_set_audio( struct hvl_voice *voice, float64 freqf )
   {
     float64 freq2;
     uint32  delta;
-    
+
     voice->vc_RingPlantPeriod = 0;
     freq2 = Period2Freq( voice->vc_RingAudioPeriod );
     delta = (uint32)(freq2 / freqf);
-    
+
     if( delta > (0x280<<16) ) delta -= (0x280<<16);
     if( delta == 0 ) delta = 1;
     voice->vc_RingDelta = delta;
   }
-  
+
   if( voice->vc_RingNewWaveform )
   {
     const int8 *src;
     uint32 i, WaveLoops;
-    
+
     src = voice->vc_RingAudioSource;
 
     WaveLoops = (1 << (5 - voice->vc_WaveLength)) * 5;
@@ -1855,13 +1661,13 @@ void hvl_play_irq( struct hvl_tune *ht )
       }
       ht->ht_GetNewPosition = 0;
     }
-    
+
     for( i=0; i<ht->ht_Channels; i++ )
       hvl_process_step( ht, &ht->ht_Voices[i] );
-    
+
     ht->ht_StepWaitFrames = ht->ht_Tempo;
   }
-  
+
   for( i=0; i<ht->ht_Channels; i++ )
     hvl_process_frame( ht, &ht->ht_Voices[i] );
 
@@ -1878,7 +1684,7 @@ void hvl_play_irq( struct hvl_tune *ht )
         ht->ht_PatternBreak = 1;
       }
     }
-    
+
     if( ht->ht_PatternBreak )
     {
       ht->ht_PatternBreak = 0;
@@ -1915,7 +1721,7 @@ void hvl_mixchunk( struct hvl_tune *ht, uint32 samples, int8 *buf1, int8 *buf2, 
 //      uint32  vu[MAX_CHANNELS];
         int32   a=0, b=0, j;
         uint32  i, chans, loops;
-  
+
   chans = ht->ht_Channels;
   for( i=0; i<chans; i++ )
   {
@@ -1925,15 +1731,15 @@ void hvl_mixchunk( struct hvl_tune *ht, uint32 samples, int8 *buf1, int8 *buf2, 
     src[i]   = ht->ht_Voices[i].vc_MixSource;
     panl[i]  = ht->ht_Voices[i].vc_PanMultLeft;
     panr[i]  = ht->ht_Voices[i].vc_PanMultRight;
-    
+
     /* Ring Modulation */
     rdelta[i]= ht->ht_Voices[i].vc_RingDelta;
     rpos[i]  = ht->ht_Voices[i].vc_RingSamplePos;
     rsrc[i]  = ht->ht_Voices[i].vc_RingMixSource;
-    
+
 //    vu[i] = 0;
   }
-  
+
   do
   {
     loops = samples;
@@ -1942,18 +1748,18 @@ void hvl_mixchunk( struct hvl_tune *ht, uint32 samples, int8 *buf1, int8 *buf2, 
       if( pos[i] >= (0x280 << 16)) pos[i] -= 0x280<<16;
       cnt = ((0x280<<16) - pos[i] - 1) / delta[i] + 1;
       if( cnt < loops ) loops = cnt;
-      
+
       if( rsrc[i] )
       {
         if( rpos[i] >= (0x280<<16)) rpos[i] -= 0x280<<16;
         cnt = ((0x280<<16) - rpos[i] - 1) / rdelta[i] + 1;
         if( cnt < loops ) loops = cnt;
       }
-      
+
     }
-    
+
     samples -= loops;
-    
+
     // Inner loop
     do
     {
@@ -1969,14 +1775,14 @@ void hvl_mixchunk( struct hvl_tune *ht, uint32 samples, int8 *buf1, int8 *buf2, 
         } else {
           j = src[i][pos[i]>>16]*vol[i];
         }
-        
+
 //        if( abs( j ) > vu[i] ) vu[i] = abs( j );
 
         a += (j * panl[i]) >> 7;
         b += (j * panr[i]) >> 7;
         pos[i] += delta[i];
       }
-      
+
       a = (a*ht->ht_mixgain)>>8;
       b = (b*ht->ht_mixgain)>>8;
 
@@ -1987,9 +1793,9 @@ void hvl_mixchunk( struct hvl_tune *ht, uint32 samples, int8 *buf1, int8 *buf2, 
 
       *(int16 *)buf1 = a;
       *(int16 *)buf2 = b;
-      
+
       loops--;
-      
+
       buf1 += bufmod;
       buf2 += bufmod;
     } while( loops > 0 );
@@ -2006,10 +1812,10 @@ void hvl_mixchunk( struct hvl_tune *ht, uint32 samples, int8 *buf1, int8 *buf2, 
 void hvl_DecodeFrame( struct hvl_tune *ht, int8 *buf1, int8 *buf2, int32 bufmod )
 {
   uint32 samples, loops;
-  
+
   samples = ht->ht_Frequency/50/ht->ht_SpeedMultiplier;
   loops   = ht->ht_SpeedMultiplier;
-  
+
   do
   {
     hvl_play_irq( ht );
