+++ Replayer_Windows/Makefile
@@ -7,6 +7,7 @@ LFLAGS += -g -lm -lwinmm
 TARGET = play_hvl.exe
 
 OBJECTS = hvl_replay.o \
+          hvl_tables.o \
           main.o
