+++ hvl2wav/hvl_tables.h
